double

+++ srcs/init_game/moves.c
@@ -5,8 +5,8 @@ void	move_left(t_data *data)
 	data->pl->dir += 0.1;
 	if (data->pl->dir >= 2 * (float)M_PI)
 		data->pl->dir -= (float)M_PI * 2;
-	data->pl->dx = cos(data->pl->dir);
-	data->pl->dy = sin(data->pl->dir);
+	data->pl->dx = cosf(data->pl->dir);
+	data->pl->dy = sinf(data->pl->dir);
 }
 
 void	move_right(t_data *data)
@@ -14,8 +14,8 @@ void	move_right(t_data *data)
 	data->pl->dir -= 0.1;
 	if (data->pl->dir < 0)
 		data->pl->dir += (float)M_PI * 2;
-	data->pl->dx = cos(data->pl->dir);
-	data->pl->dy = sin(data->pl->dir);
+	data->pl->dx = cosf(data->pl->dir);
+	data->pl->dy = sinf(data->pl->dir);
 }
 
 void	move_up(t_data *data)
